deployment/backend/requirements.txt
pandas
numpy
joblib
scikit-learn
Flask
tensorflow-cpu
gunicorn